Currency fix

+++ app/Http/Controllers/HomeController.php
@@ -3,6 +3,7 @@
 namespace App\Http\Controllers;
 
 use App\Models\User;
+use App\Models\UserCurrency;
 use App\Models\UserSearch;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
@@ -26,13 +27,19 @@ class HomeController extends Controller
      * @return \Illuminate\Contracts\Support\Renderable
      */
     public function index(){
+        /* A crutch: we should create new data with related tables */
+        if (!Auth::user()->currency) {
+            UserCurrency::init(Auth::user()->id, 0);
+            return redirect()->back();
+        }
+
         $query = UserSearch::where('user_id', '=', Auth::user()->id)->latest()->first();
 
         if ($query) {
             $recent_products = collect($query)->take(config('store.recent_products_count'));
             $products = collect(json_decode($recent_products['last_search']))->take(config('store.recent_products_count'));
 
-            $query->delete(); // clear our information when we recieve search info
+            $query->delete(); // clear our information when we receive search info
             return view('home_recent', [ 'recent' => $products ]);
         }
